remove msadaptor directory and fix dev

+++ auto_convert_llm.sh
@@ -7,6 +7,7 @@ if [ $? -ne 0 ]; then
     echo "Error:git clone MindSpeed-LLM failed"
     exit 1
 fi
+rm -rf MindSpeed-LLM/tests
 echo "------------------------------------done MindSpeed-LLM"
 
 rm -rf MindSpeed/
@@ -17,6 +18,7 @@ if [ $? -ne 0 ]; then
 fi
 cd MindSpeed
 git checkout 969686ff
+rm -rf tests_extend
 cd ..
 echo "...............................................done apply MindSpeed"
 
@@ -26,7 +28,18 @@ if [ $? -ne 0 ]; then
     echo "Error:git clone Megatron-LM failed"
     exit 1
 fi
+rm -rf Megatron-LM/tests
 echo "------------------------------------done Megatron-LM"
+
+#msadapter
+rm -rf msadapter
+git clone https://gitee.com/mindspore/msadapter.git -b master
+if [ $? -ne 0 ]; then
+    echo "Error: git clone msadapter"
+    exit 1
+fi
+rm -rf msadapter/tests
+echo "..............................................done msadapter"
 
 rm -rf transformers/
 git clone -b v4.47.0 https://github.com/huggingface/transformers.git --depth=1
@@ -34,6 +47,7 @@ if [ $? -ne 0 ]; then
     echo "Error:git clone transformers failed"
     exit 1
 fi
+rm -rf transformers/tests
 echo "------------------------------------done transformers"
 
 rm -rf peft/
@@ -51,7 +65,7 @@ python tools/transfer.py --megatron_path Megatron-LM/ --mindspeed_path MindSpeed
 --transformers_path transformers/ --mindspeed_type LLM
 
 MindSpeed_Core_MS_PATH=$(pwd)
-export PYTHONPATH=${MindSpeed_Core_MS_PATH}/msadaptor:${MindSpeed_Core_MS_PATH}/Megatron-LM/:${MindSpeed_Core_MS_PATH}/MindSpeed/:${MindSpeed_Core_MS_PATH}/transformers/src/:$PYTHONPATH
+export PYTHONPATH=${MindSpeed_Core_MS_PATH}/msadapter/mindtorch:${MindSpeed_Core_MS_PATH}/Megatron-LM/:${MindSpeed_Core_MS_PATH}/MindSpeed/:${MindSpeed_Core_MS_PATH}/transformers/src/:$PYTHONPATH
 export PYTHONPATH=${MindSpeed_Core_MS_PATH}/peft/src/:$PYTHONPATH
 echo "..............................................done set PYTHONPATH"
